# api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "info": "API methods for books and their authors",
        "version": "1.0",
        "contact": {
            "name": "API Support",
            "email": "filip.ljubic@fer.hr"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/yingray/api-book/master/LICENSE"
        }
    },
    "paths": {
        "/books/": {
            "get": {
                "summary": "Get all books from collection books",
                "description": "Returns all books",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Fetch success",
                        "schema": {
                            "$ref": "#/items/anyOf/0/properties"
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/books/{id}/": {
            "get": {
                "summary": "Get book from collection books with specified ID",
                "description": "Returns specified book",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Fetch success",
                        "schema": {
                            "$ref": "#/items/anyOf/0/properties"
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/books/{id}/authors": {
            "get": {
                "summary": "Get authors of a specified book",
                "description": "Returns specified books authors",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Fetch success",
                        "schema": {
                            "$ref": "#/items/anyOf/0/properties"
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/authors/": {
            "get": {
                "summary": "Get all authors of collection authors",
                "description": "Returns all authors",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Fetch success",
                        "schema": {
                            "$ref": "#/items/anyOf/0/properties/authors"
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "summary": "Add an author to collection authors",
                "description": "Returns added author",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Add author success",
                        "schema": {
                            "$ref": "#/items/anyOf/0/properties/authors"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/authors/{id}/": {
            "get": {
                "summary": "Get specified author of collection authors",
                "description": "Returns specified author",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Fetch success",
                        "schema": {
                            "$ref": "#/items/anyOf/0/properties/authors"
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "summary": "Updates an author from collection authors",
                "description": "Returns updated author",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Add author success",
                        "schema": {
                            "$ref": "#/items/anyOf/0/properties/authors"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "summary": "Deletes an author from collection authors",
                "description": "Returns status of deletion",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Add author success"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/openapi/": {
            "get": {
                "summary": "Get api specification",
                "description": "Returns this file",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Fetch success"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}